+++ packages/plugin-eslint/src/lib/nx/index.ts
@@ -1,2 +1,7 @@
-export { eslintConfigFromNxProjects } from './find-all-projects';
-export { eslintConfigFromNxProject } from './find-project-with-deps';
+export {
+  // eslint-disable-next-line deprecation/deprecation
+  eslintConfigFromNxProjects,
+  eslintConfigFromAllNxProjects,
+} from './find-all-projects';
+export { eslintConfigFromNxProject } from './find-project-without-deps';
+export { eslintConfigFromNxProjectAndDeps } from './find-project-with-deps';