+++ packages/plugin-eslint/src/lib/nx/projects-to-config.unit.test.ts
@@ -68,15 +68,12 @@ describe('nxProjectsToConfig', () => {
 
     expect(config).toEqual<ESLintPluginConfig>([
       {
-        eslintrc: './apps/client/.eslintrc.json',
         patterns: expect.arrayContaining(['apps/client/**/*.ts']),
       },
       {
-        eslintrc: './apps/server/.eslintrc.json',
         patterns: expect.arrayContaining(['apps/server/**/*.ts']),
       },
       {
-        eslintrc: './libs/models/.eslintrc.json',
         patterns: expect.arrayContaining(['libs/models/**/*.ts']),
       },
     ]);
@@ -108,7 +105,6 @@ describe('nxProjectsToConfig', () => {
 
     expect(config).toEqual<ESLintPluginConfig>([
       {
-        eslintrc: './libs/models/.eslintrc.json',
         patterns: expect.arrayContaining(['libs/models/**/*.ts']),
       },
     ]);
@@ -150,11 +146,9 @@ describe('nxProjectsToConfig', () => {
 
     expect(config).toEqual<ESLintPluginConfig>([
       {
-        eslintrc: './apps/client/.eslintrc.json',
         patterns: expect.arrayContaining(['apps/client/**/*.ts']),
       },
       {
-        eslintrc: './apps/server/.eslintrc.json',
         patterns: expect.arrayContaining(['apps/server/**/*.ts']),
       },
     ]);
@@ -222,11 +216,9 @@ describe('nxProjectsToConfig', () => {
           'apps/client/**/*.ts',
           'apps/client/**/*.html',
         ]),
-        eslintrc: './apps/client/.eslintrc.json',
       },
       {
         patterns: expect.arrayContaining(['apps/server/**/*.ts']),
-        eslintrc: './apps/server/.eslintrc.json',
       },
     ] satisfies ESLintPluginConfig);
   });